project: Remove project type target parsers from this project
I'm currently working on a new project management library (projector)
and intend to move these definitions into a compatibility layer their
instead of maintaining them here. This will make it easier to use newer
build frameworks with project definitions AND make this package function
more as a utility then as a feature provider for the myriad of available
build frameworks.